Access and Permissions

+++ papermerge/core/views/nodes.py
@@ -3,7 +3,8 @@ import logging
 
 from django.http import (
     HttpResponse,
-    HttpResponseBadRequest
+    HttpResponseBadRequest,
+    HttpResponseForbidden
 )
 from django.contrib.auth.decorators import login_required
 from django.urls import reverse
@@ -36,6 +37,24 @@ def browse_view(request, parent_id=None):
     for node in nodes:
         if request.user.has_perm(Access.PERM_READ, node):
             node_dict = node.to_dict()
+
+            node_dict['user_perms'] = {
+                'read': request.user.has_perm(
+                    Access.PERM_READ, node
+                ),
+                'write': request.user.has_perm(
+                    Access.PERM_WRITE, node
+                ),
+                'delete': request.user.has_perm(
+                    Access.PERM_DELETE, node
+                ),
+                'change_perm': request.user.has_perm(
+                    Access.PERM_CHANGE_PERM, node
+                ),
+                'take_ownership': request.user.has_perm(
+                    Access.PERM_TAKE_OWNERSHIP, node
+                ),
+            }
 
             if node.is_document():
                 node_dict['img_src'] = reverse(
@@ -121,8 +140,17 @@ def node_view(request, node_id):
         )
 
     if request.method == "DELETE":
-
-        node.delete()
+        if request.user.has_perm(Access.PERM_DELETE, node):
+            node.delete()
+        else:
+            msg = f"{request.user.username} does not have" +\
+                f" permission to delete {node.title}"
+            return HttpResponseForbidden(
+                json.dumps({
+                    'msg': msg
+                }),
+                content_type="application/json"
+            )
 
         return HttpResponse(
             json.dumps({
@@ -141,12 +169,28 @@ def node_view(request, node_id):
 
 @login_required
 def nodes_view(request):
+
     if request.method == "POST":
 
         data = json.loads(request.body)
         node_ids = [item['id'] for item in data]
 
         queryset = BaseTreeNode.objects.filter(id__in=node_ids)
+        for node in queryset:
+            # is used allowd to delete ?
+            if not request.user.has_perm(Access.PERM_DELETE, node):
+                # if user does not have delete permission on
+                # one node - forbid entire operation!
+                msg = f"{request.user.username} does not have" +\
+                    f" permission to delete {node.title}"
+                return HttpResponseForbidden(
+                    json.dumps({
+                        'msg': msg
+                    }),
+                    content_type="application/json"
+                )
+        # yes, user is allowed to delete all nodes,
+        # proceed with delete opration
         recursive_delete(queryset)
 
         return HttpResponse(
